backend/requirements.txt
# Core Framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0

# LangChain & AI
langchain>=0.3.0,<1.0.0
langgraph>=0.2.0,<1.0.0
langchain-google-genai>=2.0.0,<3.0.0
langchain-mongodb>=0.3.0,<1.0.0
langchain-chroma>=0.2.0,<1.0.0

# Database Clients
motor>=3.6.0,<4.0.0
chromadb>=0.5.0,<1.0.0
pymongo>=4.9.0,<5.0.0

# Google Cloud (Voice)
google-cloud-speech>=2.29.0,<3.0.0
google-cloud-texttospeech>=2.24.0,<3.0.0

# Web Search
tavily-python>=0.5.0,<1.0.0

# Configuration & Validation
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
pyyaml>=6.0.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
httpx>=0.27.0,<1.0.0
